backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
pydantic-settings==2.8.1
email-validator==2.2.0
python-jose[cryptography]==3.4.0
passlib[argon2]==1.7.4
python-multipart==0.0.20
slowapi==0.1.9
psycopg[binary]==3.2.13
httpx==0.28.1
apscheduler==3.10.4
